requirements.txt
fastapi
langchain
uvicorn
google-auth
google-auth-oauthlib
google-api-python-client
requests
pydantic
pydantic-settings
